@@ -10,20 +10,20 @@ public sealed class BundleDefinition
     /// <summary>
     /// Virtual path for the bundle.
     /// </summary>
-    public string Path { get; set; }
+    public string? Path { get; set; }
 
     /// <summary>
     /// CDN path for the bundle.
     /// </summary>
-    public string CdnPath { get; set; }
+    public string? CdnPath { get; set; }
 
     /// <summary>
     /// CDN fallback expression for the bundle.
     /// </summary>
-    public string CdnFallbackExpression { get; set; }
+    public string? CdnFallbackExpression { get; set; }
 
     /// <summary>
     /// Files to be included in the bundle.
     /// </summary>
-    public IList<string> Includes { get; internal set; }
+    public IList<string> Includes { get; internal set; } = [];
 }
