@@ -20,7 +20,7 @@ public sealed class RouteDataAttribute : ValueProviderSourceAttribute {
 
     public override IValueProvider GetValueProvider(ModelBindingExecutionContext modelBindingExecutionContext) {
         if (modelBindingExecutionContext == null) {
-            throw new ArgumentNullException("modelBindingExecutionContext");
+            throw new ArgumentNullException(nameof(modelBindingExecutionContext));
         }
 
         return new RouteDataValueProvider(modelBindingExecutionContext);
