@@ -10,8 +10,8 @@ namespace System.Web.UI.WebControls;
 /// </devdoc>
 [EditorBrowsable(EditorBrowsableState.Never)]
 public sealed class AutoGeneratedField : BoundField {
-    
-    private bool _suppressPropertyThrows = false;
+
+    private bool _suppressPropertyThrows; 
     private bool _useCheckBox;
     private bool _useCheckBoxValid;
     
@@ -22,7 +22,6 @@ public sealed class AutoGeneratedField : BoundField {
     public AutoGeneratedField(string dataField) {
         DataField = dataField;
     }
-
 
     /// <devdoc>
     /// <para>Gets or sets the display format of data in this
@@ -39,7 +38,6 @@ public sealed class AutoGeneratedField : BoundField {
         }
     }
 
-
     /// <devdoc>
     /// <para>Gets or sets the type of data being displayed by the field.</para>
     /// </devdoc>
@@ -55,7 +53,6 @@ public sealed class AutoGeneratedField : BoundField {
             ViewState["DataType"] = value;
         }
     }
-
 
     /// <devdoc>
     /// <para>Gets or sets whether the field is visible in Insert mode.  Turn off for auto-gen'd db fields</para>
@@ -114,7 +111,6 @@ public sealed class AutoGeneratedField : BoundField {
         return new AutoGeneratedField(DataField);
     }
 
-
     /// <devdoc>
     /// Returns a value to be used for design-time rendering
     /// </devdoc>
@@ -162,7 +158,6 @@ public sealed class AutoGeneratedField : BoundField {
 
     }
 
-
     protected override void InitializeDataCell(DataControlFieldCell cell, DataControlRowState rowState) {
         if (UseCheckBox) {
             CheckBox childControl = null;
@@ -197,14 +192,12 @@ public sealed class AutoGeneratedField : BoundField {
         }
     }
 
-
     /// <devdoc>
     /// </devdoc>
     protected override void OnDataBindField(object sender, EventArgs e) {
         if (UseCheckBox) {
             Control boundControl = (Control)sender;
             Control controlContainer = boundControl.NamingContainer;
-
 
             object data = GetValue(controlContainer);
  
