@@ -13,14 +13,12 @@ public class GridViewPageEventArgs : CancelEventArgs {
 
     private int _newPageIndex;
 
-
     /// <devdoc>
     /// <para>Initializes a new instance of the <see cref='System.Web.UI.WebControls.GridViewPageEventArgs'/> class.</para>
     /// </devdoc>
     public GridViewPageEventArgs(int newPageIndex) {
         this._newPageIndex = newPageIndex;
     }
-
 
     /// <devdoc>
     /// <para>Gets the index of the first new page to be displayed in the <see cref='System.Web.UI.WebControls.GridView'/>. 
@@ -32,7 +30,7 @@ public class GridViewPageEventArgs : CancelEventArgs {
         }
         set {
             if (value < 0) {
-                throw new ArgumentOutOfRangeException("value");
+                throw new ArgumentOutOfRangeException(nameof(value));
             }
             _newPageIndex = value;
         }
